remove the Flow form the discover sources and rename the files
