+++ app/src/main/java/dev/bruno/wheretowatch/features/discover/movies/CollectionMovieSource.kt
@@ -6,21 +6,12 @@ import dev.bruno.wheretowatch.features.discover.DiscoverMovieItem
 import dev.bruno.wheretowatch.services.discover.DiscoverCategory
 import dev.bruno.wheretowatch.services.discover.DiscoverContentSupplier
 import dev.bruno.wheretowatch.services.discover.MovieCollection
-import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 
-typealias MovieCollectionMap = Map<MovieCollection, ImmutableList<DiscoverMovieItem>>
-
-class CollectionMovieFlowSource @Inject constructor(
+class CollectionMovieSource @Inject constructor(
     private val supplier: DiscoverContentSupplier,
 ) {
-
-    private val state = MutableStateFlow<MovieCollectionMap>(mapOf())
-    val flow: Flow<MovieCollectionMap> = state.asStateFlow()
 
     suspend fun get(collection: MovieCollection): DiscoverContent {
         val items = supplier.get(DiscoverCategory.Collection(collection)).map { item ->
