+++ app/src/main/java/dev/bruno/wheretowatch/features/discover/movies/TopRatedSource.kt
@@ -1,25 +1,18 @@
 package dev.bruno.wheretowatch.features.discover.movies
 
+import dev.bruno.wheretowatch.features.discover.ContentList
 import dev.bruno.wheretowatch.features.discover.DiscoverContent
 import dev.bruno.wheretowatch.features.discover.DiscoverMovieItem
-import dev.bruno.wheretowatch.features.discover.DiscoverTrending
 import dev.bruno.wheretowatch.services.discover.DiscoverCategory
 import dev.bruno.wheretowatch.services.discover.DiscoverContentSupplier
-import dev.bruno.wheretowatch.services.discover.TrendWindow
 import kotlinx.collections.immutable.toImmutableList
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 
-class TrendingMovieFlowSource @Inject constructor(
+class TopRatedSource @Inject constructor(
     private val supplier: DiscoverContentSupplier,
 ) {
-    private val state = MutableStateFlow(DiscoverTrending())
-    val flow: Flow<DiscoverTrending> = state.asStateFlow()
-
-    suspend fun get(window: TrendWindow): DiscoverContent {
-        val trendingItem = supplier.get(DiscoverCategory.Trending(window)).map { item ->
+    suspend fun getTopRatedV2(): DiscoverContent {
+        val items = supplier.get(DiscoverCategory.TopRated).map { item ->
             DiscoverMovieItem(
                 id = item.id,
                 title = item.title,
@@ -31,6 +24,6 @@ class TrendingMovieFlowSource @Inject constructor(
             )
         }.toImmutableList()
 
-        return DiscoverTrending(trendWindow = window, items = trendingItem)
+        return ContentList(items)
     }
 }
