+++ app/src/main/java/dev/bruno/wheretowatch/features/discover/movies/TrendingMovieSource.kt
@@ -1,28 +1,19 @@
 package dev.bruno.wheretowatch.features.discover.movies
 
-import dev.bruno.wheretowatch.features.discover.ContentList
 import dev.bruno.wheretowatch.features.discover.DiscoverContent
 import dev.bruno.wheretowatch.features.discover.DiscoverMovieItem
+import dev.bruno.wheretowatch.features.discover.DiscoverTrending
 import dev.bruno.wheretowatch.services.discover.DiscoverCategory
 import dev.bruno.wheretowatch.services.discover.DiscoverContentSupplier
-import dev.bruno.wheretowatch.services.discover.MovieGenre
-import kotlinx.collections.immutable.ImmutableList
+import dev.bruno.wheretowatch.services.discover.TrendWindow
 import kotlinx.collections.immutable.toImmutableList
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 
-typealias PopularMap = Map<MovieGenre, ImmutableList<DiscoverMovieItem>>
-
-class PopularMovieFlowSource @Inject constructor(
+class TrendingMovieSource @Inject constructor(
     private val supplier: DiscoverContentSupplier,
 ) {
-    private val state = MutableStateFlow<PopularMap>(mapOf())
-    val flow: Flow<PopularMap> = state.asStateFlow()
-
-    suspend fun get(genre: MovieGenre = MovieGenre.ALL): DiscoverContent {
-        val items = supplier.get(DiscoverCategory.Popular(genre)).map { item ->
+    suspend fun get(window: TrendWindow): DiscoverContent {
+        val trendingItem = supplier.get(DiscoverCategory.Trending(window)).map { item ->
             DiscoverMovieItem(
                 id = item.id,
                 title = item.title,
@@ -34,6 +25,6 @@ class PopularMovieFlowSource @Inject constructor(
             )
         }.toImmutableList()
 
-        return ContentList(items)
+        return DiscoverTrending(trendWindow = window, items = trendingItem)
     }
 }
